# perf: Avoid looking up unnecessary TSDB symbols during Volume API