--- a/pkg/storage/store_test.go
+++ b/pkg/storage/store_test.go
@@ -219,7 +219,8 @@ func getLocalStore(path string, cm ClientMetrics) Store {
 					PeriodicTableConfig: config.PeriodicTableConfig{
 						Prefix: "index_",
 						Period: time.Hour * 168,
-					}},
+					},
+				},
 				RowShards: 16,
 			},
 		},
@@ -866,6 +867,10 @@ func (f fakeChunkFilterer) ShouldFilter(metric labels.Labels) bool {
 	return metric.Get("foo") == "bazz"
 }
 
+func (f fakeChunkFilterer) RequiredLabelNames() []string {
+	return []string{"foo"}
+}
+
 func Test_ChunkFilterer(t *testing.T) {
 	s := &LokiStore{
 		Store: storeFixture,
@@ -921,7 +926,6 @@ func Test_PipelineWrapper(t *testing.T) {
 	s.SetPipelineWrapper(wrapper)
 	ctx = user.InjectOrgID(context.Background(), "test-user")
 	logit, err := s.SelectLogs(ctx, logql.SelectLogParams{QueryRequest: newQuery("{foo=~\"ba.*\"}", from, from.Add(1*time.Hour), []astmapper.ShardAnnotation{{Shard: 1, Of: 5}}, nil)})
-
 	if err != nil {
 		t.Errorf("store.SelectLogs() error = %v", err)
 		return
@@ -952,7 +956,6 @@ func Test_PipelineWrapper_disabled(t *testing.T) {
 	ctx = user.InjectOrgID(context.Background(), "test-user")
 	ctx = httpreq.InjectHeader(ctx, httpreq.LokiDisablePipelineWrappersHeader, "true")
 	logit, err := s.SelectLogs(ctx, logql.SelectLogParams{QueryRequest: newQuery("{foo=~\"ba.*\"}", from, from.Add(1*time.Hour), []astmapper.ShardAnnotation{{Shard: 1, Of: 5}}, nil)})
-
 	if err != nil {
 		t.Errorf("store.SelectLogs() error = %v", err)
 		return
@@ -1292,7 +1295,8 @@ func TestStore_indexPrefixChange(t *testing.T) {
 			PeriodicTableConfig: config.PeriodicTableConfig{
 				Prefix: "index_",
 				Period: time.Hour * 24,
-			}},
+			},
+		},
 	}
 
 	schemaConfig := config.SchemaConfig{
@@ -1366,7 +1370,8 @@ func TestStore_indexPrefixChange(t *testing.T) {
 			PeriodicTableConfig: config.PeriodicTableConfig{
 				Prefix: "index_tsdb_",
 				Period: time.Hour * 24,
-			}},
+			},
+		},
 		RowShards: 2,
 	}
 	schemaConfig.Configs = append(schemaConfig.Configs, periodConfig2)
@@ -1471,7 +1476,8 @@ func TestStore_MultiPeriod(t *testing.T) {
 					PeriodicTableConfig: config.PeriodicTableConfig{
 						Prefix: "index_",
 						Period: time.Hour * 24,
-					}},
+					},
+				},
 			}
 
 			periodConfigV11 := config.PeriodConfig{
@@ -1483,7 +1489,8 @@ func TestStore_MultiPeriod(t *testing.T) {
 					PeriodicTableConfig: config.PeriodicTableConfig{
 						Prefix: "index_",
 						Period: time.Hour * 24,
-					}},
+					},
+				},
 				RowShards: 2,
 			}
 
@@ -1562,7 +1569,6 @@ func TestStore_MultiPeriod(t *testing.T) {
 			}
 		})
 	}
-
 }
 
 func mustParseLabels(s string) []logproto.SeriesIdentifier_LabelsEntry {
@@ -1829,7 +1835,8 @@ func TestStore_BoltdbTsdbSameIndexPrefix(t *testing.T) {
 					PeriodicTableConfig: config.PeriodicTableConfig{
 						Prefix: "index_",
 						Period: time.Hour * 24,
-					}},
+					},
+				},
 				RowShards: 2,
 			},
 			{
@@ -1842,7 +1849,8 @@ func TestStore_BoltdbTsdbSameIndexPrefix(t *testing.T) {
 					PeriodicTableConfig: config.PeriodicTableConfig{
 						Prefix: "index_",
 						Period: time.Hour * 24,
-					}},
+					},
+				},
 			},
 		},
 	}
@@ -1980,7 +1988,8 @@ func TestStore_SyncStopInteraction(t *testing.T) {
 					PeriodicTableConfig: config.PeriodicTableConfig{
 						Prefix: "index_",
 						Period: time.Hour * 24,
-					}},
+					},
+				},
 				RowShards: 2,
 			},
 			{
@@ -1993,7 +2002,8 @@ func TestStore_SyncStopInteraction(t *testing.T) {
 					PeriodicTableConfig: config.PeriodicTableConfig{
 						Prefix: "index_",
 						Period: time.Hour * 24,
-					}},
+					},
+				},
 			},
 		},
 	}